+++ src/Ext/House/Body.h
@@ -36,6 +36,8 @@ public:
 		BuildingClass* Factory_NavyType;
 		BuildingClass* Factory_AircraftType;
 
+		std::vector<int> AITriggers_ValidList;
+
 		//Read from INI
 		bool RepairBaseNodes[3];
 
@@ -113,6 +115,9 @@ public:
 	static int TotalHarvesterCount(HouseClass* pThis);
 	static HouseClass* GetHouseKind(OwnerHouseKind kind, bool allowRandom, HouseClass* pDefault, HouseClass* pInvoker = nullptr, HouseClass* pVictim = nullptr);
 	static CellClass* GetEnemyBaseGatherCell(HouseClass* pTargetHouse, HouseClass* pCurrentHouse, CoordStruct defaultCurrentCoords, SpeedType speedTypeZone, int extraDistance = 0);
+	static bool PrerequisitesMet(HouseClass* const pThis, TechnoTypeClass* const pItem, const std::map<TechnoTypeClass*, int> ownedBuildings, bool skipSecretLabChecks);
+	static bool HasGenericPrerequisite(int idx, std::map<TechnoTypeClass*, int> ownedBuildings);
+	static int FindGenericPrerequisite(const char* id);
 
 	static void SetSkirmishHouseName(HouseClass* pHouse);