+++ src/Ext/Rules/Body.cpp
@@ -184,6 +184,18 @@ void RulesExt::ExtData::LoadBeforeTypeData(RulesClass* pThis, CCINIClass* pINI)
 
 		this->AIScriptsLists.emplace_back(std::move(objectsList));
 	}
+
+	this->NewTeamsSelector.Read(exINI, "AI", "NewTeamsSelector");
+	this->NewTeamsSelector_SplitTriggersByCategory.Read(exINI, "AI", "NewTeamsSelector.SplitTriggersByCategory");
+	this->NewTeamsSelector_EnableFallback.Read(exINI, "AI", "NewTeamsSelector.EnableFallback");
+	this->NewTeamsSelector_MergeUnclassifiedCategoryWith.Read(exINI, "AI", "NewTeamsSelector.MergeUnclassifiedCategoryWith");
+	this->NewTeamsSelector_UnclassifiedCategoryPercentage.Read(exINI, "AI", "NewTeamsSelector.UnclassifiedCategoryPercentage");
+	this->NewTeamsSelector_GroundCategoryPercentage.Read(exINI, "AI", "NewTeamsSelector.GroundCategoryPercentage");
+	this->NewTeamsSelector_AirCategoryPercentage.Read(exINI, "AI", "NewTeamsSelector.AirCategoryPercentage");
+	this->NewTeamsSelector_NavalCategoryPercentage.Read(exINI, "AI", "NewTeamsSelector.NavalCategoryPercentage");
+
+	// Section Generic Prerequisites
+	FillDefaultPrerequisites();
 }
 
 // this runs between the before and after type data loading methods for rules ini
@@ -203,6 +215,62 @@ void RulesExt::ExtData::LoadAfterTypeData(RulesClass* pThis, CCINIClass* pINI)
 		return;
 
 	INI_EX exINI(pINI);
+}
+
+void RulesExt::FillDefaultPrerequisites()
+{
+	if (RulesExt::Global()->GenericPrerequisitesNames.Count != 0)
+		return;
+
+	CCINIClass::INI_Rules;
+	DynamicVectorClass<int> empty;
+
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("POWER"); // Official index: -1
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisitePower);
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("FACTORY"); // -2
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisiteFactory);
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("BARRACKS"); // -3
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisiteBarracks);
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("RADAR"); // -4
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisiteRadar);
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("TECH"); // -5
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisiteTech);
+	RulesExt::Global()->GenericPrerequisitesNames.AddItem("PROC"); // -6
+	RulesExt::Global()->GenericPrerequisites.AddItem(RulesClass::Instance->PrerequisiteProc);
+
+	// If [GenericPrerequisites] is present will be added after these.
+	// Also the originals can be replaced by new ones
+	int genericPreqsCount = CCINIClass::INI_Rules->GetKeyCount("GenericPrerequisites");
+
+	for (int i = 0; i < genericPreqsCount; ++i)
+	{
+		DynamicVectorClass<int> objectsList;
+		char* context = nullptr;
+		CCINIClass::INI_Rules->ReadString("GenericPrerequisites", CCINIClass::INI_Rules->GetKeyName("GenericPrerequisites", i), "", Phobos::readBuffer);
+		char* name = (char*)CCINIClass::INI_Rules->GetKeyName("GenericPrerequisites", i);
+
+		for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			int idx = BuildingTypeClass::FindIndex(cur);
+			if (idx >= 0)
+				objectsList.AddItem(idx);
+		}
+
+		int index = RulesExt::Global()->GenericPrerequisitesNames.FindItemIndex(name);
+		if (index < 7 && index > 0)
+		{
+			// Overwrites a vanilla generic prerequisite
+			RulesExt::Global()->GenericPrerequisites[index] = objectsList;
+		}
+		else
+		{
+			// New generic prerequisite
+			RulesExt::Global()->GenericPrerequisitesNames.AddItem(name);
+			RulesExt::Global()->GenericPrerequisites.AddItem(objectsList);
+		}
+
+		objectsList.Clear();
+	}
 }
 
 // =============================
@@ -282,6 +350,16 @@ void RulesExt::ExtData::Serialize(T& Stm)
 		.Process(this->Vehicles_DefaultDigitalDisplayTypes)
 		.Process(this->Aircraft_DefaultDigitalDisplayTypes)
 		.Process(this->ShowDesignatorRange)
+		.Process(this->GenericPrerequisites)
+		.Process(this->GenericPrerequisitesNames)
+		.Process(this->NewTeamsSelector)
+		.Process(this->NewTeamsSelector_SplitTriggersByCategory)
+		.Process(this->NewTeamsSelector_EnableFallback)
+		.Process(this->NewTeamsSelector_MergeUnclassifiedCategoryWith)
+		.Process(this->NewTeamsSelector_UnclassifiedCategoryPercentage)
+		.Process(this->NewTeamsSelector_GroundCategoryPercentage)
+		.Process(this->NewTeamsSelector_AirCategoryPercentage)
+		.Process(this->NewTeamsSelector_NavalCategoryPercentage)
 		.Process(this->DropPodTrailer)
 		.Process(this->PodImage)
 		;