+++ src/Ext/Team/Body.h
@@ -1,11 +1,74 @@
 #pragma once
 #include <TeamClass.h>
+#include <AITriggerTypeClass.h>
 
 #include <Helpers/Enumerators.h>
 #include <Helpers/Macro.h>
 #include <Utilities/Container.h>
 #include <Utilities/TemplateDef.h>
+#include <Ext/HouseType/Body.h>
+#include <Ext/House/Body.h>
+#include <Ext/Rules/Body.h>
+#include <Ext/TechnoType/Body.h>
 #include <Phobos.h>
+
+enum teamCategory
+{
+	None = 0, // No category. Should be default value
+	Ground = 1,
+	Air = 2,
+	Naval = 3,
+	Unclassified = 4
+};
+
+struct TriggerElementWeight
+{
+	double Weight = 0.0;
+	AITriggerTypeClass* Trigger = nullptr;
+	teamCategory Category = teamCategory::None;
+
+	//need to define a == operator so it can be used in array classes
+	bool operator==(const TriggerElementWeight& other) const
+	{
+		return (Trigger == other.Trigger && Weight == other.Weight && Category == other.Category);
+	}
+
+	//unequality
+	bool operator!=(const TriggerElementWeight& other) const
+	{
+		return (Trigger != other.Trigger || Weight != other.Weight || Category == other.Category);
+	}
+
+	bool operator<(const TriggerElementWeight& other) const
+	{
+		return (Weight < other.Weight);
+	}
+
+	bool operator<(const double other) const
+	{
+		return (Weight < other);
+	}
+
+	bool operator>(const TriggerElementWeight& other) const
+	{
+		return (Weight > other.Weight);
+	}
+
+	bool operator>(const double other) const
+	{
+		return (Weight > other);
+	}
+
+	bool operator==(const double other) const
+	{
+		return (Weight == other);
+	}
+
+	bool operator!=(const double other) const
+	{
+		return (Weight != other);
+	}
+};
 
 class TeamExt
 {
@@ -81,5 +144,13 @@ public:
 	};
 
 	static ExtContainer ExtMap;
+
+	static bool HouseOwns(AITriggerTypeClass* pThis, HouseClass* pHouse, bool allies, std::vector<TechnoTypeClass*> list);
+	static bool HouseOwnsAll(AITriggerTypeClass* pThis, HouseClass* pHouse, std::vector<TechnoTypeClass*> list);
+	static bool EnemyOwns(AITriggerTypeClass* pThis, HouseClass* pHouse, HouseClass* pEnemy, bool onlySelectedEnemy, std::vector<TechnoTypeClass*> list);
+	static bool EnemyOwnsAll(AITriggerTypeClass* pThis, HouseClass* pHouse, HouseClass* pEnemy, std::vector<TechnoTypeClass*> list);
+	static bool NeutralOwns(AITriggerTypeClass* pThis, std::vector<TechnoTypeClass*> list);
+	static bool NeutralOwnsAll(AITriggerTypeClass* pThis, std::vector<TechnoTypeClass*> list);
+	static bool CountConditionMet(AITriggerTypeClass* pThis, int nObjects);
 
 };