+++ src/Ext/TechnoType/Body.cpp
@@ -6,6 +6,7 @@
 #include <Ext/BuildingType/Body.h>
 #include <Ext/BulletType/Body.h>
 #include <Ext/Techno/Body.h>
+#include <Ext/House/Body.h>
 
 #include <Utilities/GeneralUtils.h>
 
@@ -406,6 +407,135 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 		this->InterceptorType.reset();
 	}
 
+	this->ConsideredNaval.Read(exINI, pSection, "ConsideredNaval");
+	this->ConsideredVehicle.Read(exINI, pSection, "ConsideredVehicle");
+	this->ConsideredSecretLabTech.Read(exINI, pSection, "ConsideredSecretLabTech");
+
+	// Secret.RequiredHouses contains a list of HouseTypeClass indexes
+	this->Secret_RequiredHouses.clear();
+	const char* key = "SecretLab.RequiredHouses";
+	char* context = nullptr;
+	pINI->ReadString(pSection, key, "", Phobos::readBuffer);
+
+	for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+	{
+		std::string item(cur);
+		this->Secret_RequiredHouses.push_back(item);
+	}
+
+	key = nullptr;
+
+	// Secret.ForbiddenHouses contains a list of HouseTypeClass indexes
+	this->Secret_ForbiddenHouses.clear();
+	key = "SecretLab.ForbiddenHouses";
+	context = nullptr;
+	pINI->ReadString(pSection, key, "", Phobos::readBuffer);
+
+	for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+	{
+		std::string item(cur);
+		this->Secret_ForbiddenHouses.push_back(item);
+	}
+
+	key = nullptr;
+
+	// Prerequisite.RequiredTheaters contains a list of theader names
+	this->Prerequisite_RequiredTheaters.clear();
+	key = "Prerequisite.RequiredTheaters";
+	context = nullptr;
+	pINI->ReadString(pSection, key, "", Phobos::readBuffer);
+
+	for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+	{
+		int index = Theater::FindIndex(cur);
+		if (index != -1)
+			this->Prerequisite_RequiredTheaters.push_back(index);
+	}
+
+	key = nullptr;
+
+	// Prerequisite with Generic Prerequistes support.
+	// Note: I have no idea of what could happen in all the game engine logics if I push the negative indexes of the Ares generic prerequisites directly into the original Prerequisite tag... for that reason this tag is duplicated for working with it
+	this->Prerequisite.clear();
+	key = "Prerequisite";
+	context = nullptr;
+	pINI->ReadString(pSection, key, "", Phobos::readBuffer);
+
+	for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+	{
+		int idx = TechnoTypeClass::FindIndex(cur);
+		if (idx >= 0)
+		{
+			this->Prerequisite.push_back(idx);
+		}
+		else
+		{
+			int index = HouseExt::FindGenericPrerequisite(cur);
+			if (index < 0)
+				this->Prerequisite.push_back(index);
+		}
+	}
+
+	key = nullptr;
+
+	// Prerequisite.Negative with Generic Prerequistes support
+	this->Prerequisite_Negative.clear();
+	key = "Prerequisite.Negative";
+	context = nullptr;
+	pINI->ReadString(pSection, key, "", Phobos::readBuffer);
+
+	for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+	{
+		int idx = TechnoTypeClass::FindIndex(cur);
+		if (idx >= 0)
+		{
+			this->Prerequisite_Negative.push_back(idx);
+		}
+		else
+		{
+			int index = HouseExt::FindGenericPrerequisite(cur);
+			if (index < 0)
+				this->Prerequisite_Negative.push_back(index);
+		}
+	}
+
+	key = nullptr;
+
+	// Prerequisite.ListX with Generic Prerequistes support
+	this->Prerequisite_ListVector.clear();
+	this->Prerequisite_Lists.Read(exINI, pSection, "Prerequisite.Lists");
+
+	if (Prerequisite_Lists.Get() > 0)
+	{
+		for (int i = 1; i <= Prerequisite_Lists.Get(); i++)
+		{
+			char keySection[32];
+			_snprintf_s(keySection, sizeof(keySection), "Prerequisite.List%d", i);
+
+			DynamicVectorClass<int> objectsList;
+			char* context2 = nullptr;
+			pINI->ReadString(pSection, keySection, "", Phobos::readBuffer);
+
+			for (char* cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context2); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context2))
+			{
+				int idx = TechnoTypeClass::FindIndex(cur);
+				if (idx >= 0)
+				{
+					objectsList.AddItem(idx);
+				}
+				else
+				{
+					int index = HouseExt::FindGenericPrerequisite(cur);
+					if (index < 0)
+						objectsList.AddItem(index);
+				}
+			}
+
+			this->Prerequisite_ListVector.push_back(objectsList);
+			objectsList.Clear();
+		}
+	}
+
 	if (this->OwnerObject()->WhatAmI() == AbstractType::InfantryType)
 	{
 		if (this->DroppodType == nullptr)
@@ -547,6 +677,17 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->IronCurtain_KillWarhead)
 
 		.Process(this->Explodes_KillPassengers)
+		.Process(this->Prerequisite_RequiredTheaters)
+		.Process(this->Prerequisite)
+		.Process(this->Prerequisite_Negative)
+		.Process(this->Prerequisite_Lists)
+		.Process(this->Prerequisite_ListVector)
+		.Process(this->ConsideredNaval)
+		.Process(this->ConsideredVehicle)
+		.Process(this->ConsideredSecretLabTech)
+		.Process(this->Secret_RequiredHouses)
+		.Process(this->Secret_ForbiddenHouses)
+
 		.Process(this->DeployFireWeapon)
 		.Process(this->TargetZoneScanType)